Modif route delete

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require("../models/Thing");
+const fs = require("fs");
 
 exports.createThing = (req, res, next) => {
 
@@ -29,7 +30,7 @@ exports.modifyThing = (req, res, next) => {
     Thing.findOne({_id :req.params.id})
     .then((thing) => {
         if(thing.userId !== req.auth.userId) {
-            res.status(401).json({ message : "Non autorisé" })
+            res.status(401).json({ message : "Non autorisé" });
         } else {
             Thing.updateOne({ _id: req.params.id }, {...thingObject, _id: req.params.id})
             .then(() => res.status(200).json({message: "Modfication effectuée !"}))
@@ -42,9 +43,20 @@ exports.modifyThing = (req, res, next) => {
 
 exports.deleteThing = (req, res, next) => {
 
-    Thing.deleteOne({ _id: req.params.id})
-    .then(() => res.status(200).json({ message: "Objet supprimé !" }))
-    .catch(() => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+        if( thing.userId !== req.auth.userId){
+            res.status(401).json({ message : "Non autorisé" });
+        } else {
+            const filename = thing.imageUrl.split("/images/")[1];
+            fs.unlink(`images/${filename}`, () => {
+                Thing.deleteOne({_id: req.params.id})
+                .then(() => res.status(200).json({ message: "Objet supprimé !"}))
+                .catch(error => res.status(401).json({ error }));
+            });
+        }
+    })
+    .catch(error => res.status(500).json({ error }));
 
 }
